Refactor, make bullseye menu file (e.g. bullseye.asc) optional
Use console.uselect() to list the bulletin files, if the menu file
(e.g. bullseye.asc) doesn't exist. For Amessyroom.

Don't replace the Pause (hit a key) prompt, since that messes up
localizaton support (reverts to the text.dat Pause prompt, instead of the
text.*.ini translated Pause prompt). The replace/revert text buffers should
really be a stack for a full fix for this.

Missing bulletin files are automatically excluded from the bulletin list.

The rest is just clean-up from the conversion from Baja.

This is still a very old/basic script, but at least it's not quite as hideous
and outdated as before.

Increment the version to 3.00.