Allow time format to be set via modopts.ini [logonlist] time_fmt
... in strftime format (https://cplusplus.com/reference/ctime/strftime/)

If you (sysop) have last_few_callers_fmt set to a custom value in your
modopts.ini, you'll need to update that format as a result of this change:
the number and type of arguments have now changed: instead of the logon
time be passed as hours and minutes (integers), it's now passed as a single
a string. That string is the result of a call to strftime() with the time_fmt
or (by default), "%H:%M".

This change was made to accommodate request by NeoArata (THEREALM)