Make the [BBS] MaxSessionInactivity 10m (not 0), when not set in sbbs.ini
We've had this value as the default in sbbs.ini for a while now, so should be
safe to use as the upgrading-sysop (who wouldn't have this value in their
sbbs.ini file, most likely) setting.

Not all default sbbs.ini values are true "defaults" (meaning, if they don't
exist in the file, then same value would be used) - but we should do an audit
and correct those mismatches at some point.