Don't display menu of file xfer prots when auto-selecting user's default prot
When a user has a "default download protocol" selected (configured for their
user account), don't display a menu of file transfer protocols when we're
just going to auto-select their default anyway (e.g. when downloading a QWK
packet).

This change also introduces an argument (%s, the protocol name) in the
StartXferNow text.dat string, so that if/when the user forgets which default
download transfer protocol they had selected, they'll be reminded ("oh yeah,
I need start an XMODEM download!").

This fixes issue #767