Save/reuse the mail save path/filename as a user property (in data/user/*.ini)
This is a fix for issue #340

Another idea would be to store a history of mail save paths and allow the user
to scroll through them/choose with the up and down arrow keys, but that goes
beyond the original feature request. So this just stores/reuses the last
successfully used path/filename.