A faster, better looking, less janky version of the FTN nodelist browser. Should be feature-equivalent with the old one, though the navigation is a bit different.
If you don't like it, dig the old one out of git and run it from
some other location.

This is a build of:
https://gitlab.synchro.net/echicken/nodelist-browser
and it will be maintained there for now. I'll try to remember to
update this copy as needed. Do not edit / commit changes to this
file; it's transpiled from TypeScript and not meant to be edited
by hand. Do not read this file if you value your sanity.