Don't use QWK reply message date unless time zone is also specified
For regular user QWK Relpy packet uploads only, if no timezone is specified
(e.g. via @TZ kludge or HEADERS.DAT), then over-ride the message's "posted"
date/time with the current date/time since we're going to set the message's
timezone to the BBS's local timezone as well.

This is a fix for issue #783 reported by Chris Jacobs.

If/when we support user-specified timezones, then this likely would be a
place where we'd want to use the user's timezone.