Track broker-connected status via Mosquitto connect/disconnect callbacks
I think this might fix issue #781. I suspect that SBBS (the MQTT client) is
being disconnected by the server ("due to protocol error") *after* the
call to mosquitto_connect_bind() is successful. We don't have any correponding
log output for this case, but at least we can track the connection status
accurately using the Mosquitto client callbacks and not try to publish when
we're not connected.