Split-up the opening and writing of msg pointer files, to help debug issue
Reported by Keyop (upon being auto-disconnected overnight):
!ERROR 9 (Bad file descriptor) in data_ovl.cpp line 47 (putmsgptrs) writing
"message pointers" access=0

Unfortunately, the way putmsgptrs() was written, we couldn't tell if this was
an open (data/user/*.subs file) issue or a writing issue.

So create putmsgptrs_fp() and use that from sbbs_t::putmsgptrs(), so we can
log a different error (with more accurate details) if it's a file-open failure
versues a file-write failure.