Close any open message base after input timeout (5 minutes)
As Nelgin pointed out, the nntpservice can leave a message bases open while
a client is idle. Normally, an inactive client would be disconnected after 5
minutes of inactivity, so that wouldn't be much of an issue. However, if
authenticated as a user with the H-exemption, no auto-disconnection after
input timeout would happen. So let's close any open message base for good
measure.

The input/receive timeout probably should be configurable and lowered to
a smaller duration (one minute?). I'll leave that for another commit.