Display horizontal splitter in the correct spot.
For who knows how long now, the horizontal splitter that allows the sysop
to control the size of the top 2 panels (server forms), was displayed above
the top panel, not between the top and bottom panel, thus making the panel
heights not controllable. Weird. Change the order of making the controls
visible to insure that the splitter is displayed below the top panel and
on top of the bottom panel.

Also (non-functional changes):
Remove redunant mutex from client_on() - makes no difference.
Use try/catch in client_on() to try to catch VCL exceptions - doesn't catch
the "invalid index" error popups that we see on occasion. <shrug>