Change the height of the horizontal splitter to the minimum (1 pixel)
I don't know why this was set so wide (10 pixels), but that did help to easily
see that it was being placed in the wrong spot in the form (between the tool
button bar and the top panel). But we don't need it to be that thick. 1 pixel
is still enough to be able to grab it with the mouse and resize the top/bottom
panels.