Don't allow FileBase.update() to create a duplicate file
FileBase.update() will now throw an exception if attempting to rename a file
to a filename that already exists in the filebase index.

Prior to this change, one could (via JS methods) rename a file in the base to
create a duplicate filename which would corrupt the base (one index entry and
two header records for the same filename).