Use JS_ValueToECMAUint32() instead of JS_ValueToInt32() for scan_ptr prop set
scan_cfg and last_read props too. This gives us use of the entire 32-bit range
of an unsigned int. But does someone actually have a message base with message
numbers in the billions?

This is an attempt to fix isssue #792, but I suspect scan_ptr might instead be
trying to be set to NaN or ... ?