Fix config test for va_copy() on darwin.
Because the appropriate header wasn't included, there was no
prototype.  Not having a prototype is an error in C99, which caused
the test to indicate that there was no va_copy() on darwin.

This in turn caused jsprf.cpp to try to copy a va_list using
assignment.

This is invalid, so the compiler would normally return an error.

Somwhere in here, that error became a clang crash (not trivially
reproducable).

At the end of the day, ancient broken tests in config caused the
darwin Spidermonkey builds to fail.  The new patch fixes up the
configure tests, and darwin Synchronet builds are re-enabled.