Fix reversed mouse wheel on Win32GDI...
From MSDN:
A positive value indicates that the wheel was rotated forward
away from the user; a negative value indicates that the wheel
was rotated backward, toward the user.

Fixes bug 143.

However, this re-opens the confusion around request 75...
nigel and DigitalMan should take a look at this more closely
in rc2 (or an updated nightly).