Have the IC focus follow the X11 focus
Basically, If we get a FocusOut, call XUnsetICFocus(), and if we
get a FocusIn, call XSetICFocus().  Since we're doing this, don't
call XSetICFocus() unconditionally at start, wait for the server
to send us the initial FocusIn instead.

The only bit I'm not sure of is if I got the times to ignor
messages right... see here:
https://tronche.com/gui/x/xlib/events/input-focus/normal-and-grabbed.html
If you want to check my work.

May fix the focus follows mouse issue reported by Cyan.