More graphical fixes...
Force mouse support in graphics mode even if user disabled
it for the BBS.  It would be nice if mouse support in the menus
was a separate option than "terminal doesn't support mouse",
and we know the terminal supports the mouse when we detect
graphics mode.

Move top/margin/width calculations into separate functions.
For level 1, center the playboard.

Fix an apparent off-by-one drawing the title bar.