Start working out how to build cryptlib using cmake.
The tricky part is really extracting the source and applying the
patches THEN configuring cryptlib.

I can do this with ExternalProject, but I'm not sure exactly how
to hook this up yet.

This also requires patch and perl and a supported md5 thing to be
installed in a place CMake will find them, which sill be
challenging for Windows.  If Windows (or compilers for Windows)
have started shipping with git, I can use git apply instead of
patch, but the rest is still kinda fuzzy.

This does at least fix up the terrible serialized build though,
so once I get this working, it will work well with values other
than -j1.