Improve input in GDI mode
This may be a fix for bug 150 (which was reported against SDL mode).
Basically, Windows will report AltGr as Alt + Ctrl, so we need to
be able to parse that.

Since we're here, support WM_UNICHAR to avoid the need to have
Windows convert to "ANSI" for us, and allow direct unicode input
(even in "ANSI" mode).