Fix macOS bug where the *first* entry on a new system could not be added
When the code to use URLForDirectory to get paths was written, it
didn't create the paths, so the files in them could not be created
either.

If SyncTERM had ever ran using the old FSFindFolder() method, the
path would already exist, so things would work.

Also, since we're here, use mkpath() instead of MKDIR().