Ensure graphics are cached successfully.
It appears that cache operations in SyncTERM on Windows does not
currently work with subdirectories for some reason (likely
backslashes).  Ensure that after sending a file to be cached, it
correctly appears in the cache, and fall back to not using graphics.

This will fix the blank board issue in minesweeper with a Windows
SyncTERM and fallback to text mode.

Once SyncTERM is fixed, it should work in graphics mode too.