Fix cache subdirectories on Windows.
Windows appears to treat backslashes and slashes the same when
using the API, but Windows always returns backslashes.  Convert
backslashes to slashes in clean_path() to ensure paths compare the
same regardless of them being specified with backslashes or slashes.

There's still an open question of what to do about paths received
from the remote with backslashes in them, but I suspect that doing
that would actually work on *nix systems and break on Windows systems,
so hopefully whoever does that can fogure out what's going on.