More changes around fullscreen/maximized/etc for SDL
On macOS, fullscreen and maximized are very slightly different, so
don't force one to mean the other.  However, when entering or
leaving fullscreen mode, we need to explicitly make the window
resizable or not.

Also, update the uifc screen title when the mode changes and we're
at one of the main menus... still a bit icky, but likely the best
that can be done (won't update if you toggle fullscreen from a
sub-menu for example).