Rejigger width iterpolation.
The old code expected the L1 cache to be fairly large, and the
prefetcher to be fairly smart, and did updates by columns to save
some math.  This change performs width interpolation row-by-row
so even the dumbest prefetcher can get it right, and there's no
need to keep the whole source and destination images in the cache.

This may help out older processors when scaling with interpolation
(most commonly used in fullscreen).

It's entirely possible though that this won't be enough and they'll
still need to use "External" scaling.