Pass the ctrl directory on the SCFG command-line
Fixes issue were jsexec was used to invoke this script and a ctrl dir
(different than the contents of the SBBSCTRL env var) was passed on the
command-line (i.e. testing purpose). Use the correct ctrl directory.

Create the data/text directory if it doesn't already exist. This dir
is normally created whenever SBBS is run, so it should normally already
exist, but if someone runs this script via jsexec before (ever) running
sbbs, then it's possible the data/text directory does not yet exist.