Deal with more windows console braindeadedness.
The window and the screen buffer are two different things... the
screen buffer must always be the same size as or larger than the
window.  Further, there are times where it's not possible to resize
the windown programatically (new terminal for example).

There still appears to be a stupid bug in what I assume is the
Windows Terminal LCF flag implementation which causes writes to
the beginning of a line that occur after a write to the end of the
previous line to be in the wrong position.  If it's not the LCF flag
(which isn't turned on), it's in their wrapping thing which is extra
irritating since I explocitly turn that off.