Set the size of the CONSOLE_SCREEN_BUFFER_INFOEX first
This allows GetConsoleScreenBufferInfoEx() to work for palette
changes, allowing Console to not suck.  However, we now prefer
ANSI mode over Win32 Console, so make that so it can actually
fail on Windows of stdout is a TTY and we can't set the
ENABLE_VIRTUAL_TERMINAL_PROCESSING flag.

With this, ANSI mode is used in Windows Terminal, and Win32 Console
is used in Legacy Console, and all should be good with the world.