Fix an off-by-less-than-one issue in bitmap_double_mult_inside()
This was sometimes causing an integer scaled mode (ie: 4×) to be
reduced to very slightly less than that (ie: 3.996255×), making the
window one pixel smaller than it was supposed to be due to integer
truncation.

This could result in being unable to increase the window size using
the Alt+Right-Arrow shortcut and could make new windows one full
size smaller than they needed to be (such as C64 windows).

While we're here, fix the SDL and X11 outputs to trust what
bitmap_drv_init_mode() does.  It was fixed when GDI mode was being
written, but the other drivers weren't updated to take advantage of
that.

Should resolve issue 156 reported by DigitalMan