DDMsgReader: Sorting of sub-boards when changing to another sub-board (for nelgin). Toggleable behavior to only show sub-boards with new messages in the indexed-mode newscan (for kdi4hw). Internal: Updated for bbs.msg_number and bbs.smb_curmsg being writeable in Synchronet 3.20.
The new configuration option subBoardChangeSorting specifies the sub-board sorting (None, Alphabetical, LatestMsgDateOldestFirst, or LatestMsgDateNewestFirst). This is a default for a user option that users can change for themselves.

The new configuration option indexedModeNewscanOnlyShowSubsWithNewMsgs sets the default setting for this behavior (true or false).

Updated for bbs.msg_number and bbs.smb_curmsg being writeable in Synchronet 3.20 - If these can be changed, there's no need to write the drop file DDML_SyncSMBInfo.txt containing information about the message being replied to etc. for message editors to use