sbbs_t::protnum() and protname() now accept the transfer type as argument
It's possible to have multiple transfer protocols with the same mmemonic, but
with different transfer types supports (which would be weird, but possible).
This allows us to use protnum() to replace a lot of copy/pasted prot looping
and comparing logic throughout sbbs.

Also added a variant of sbbs_t::quit_key() that takes and returns a string
(for easy concatenation to strings of key chars).