When RIP changes the vstat mode, also change the cursor
When changing between 8 and 16 row fonts in RIP mode, the cursor
start and end was left for the old font size... 16 -> 8 would leave
the cursor one line below the current position, and 8 -> 16 would
leave the cursor in the middle of the cell.

This hacks the vstat deeper to fix up the cursor as well.

Reported by skipperdoodle1947.