Support string (text.dat string ID) arguments to bbs.text() and system.text()
Currently, this is a slower way to look-up a text.dat string, but requires no
load/require of text.js and allows for a shorter syntax to get a text.dat
string, i.e. bbs.text("Quit") instead of bbs.text(bbs.text.Quit);

Although the bbs.text() ID lookup is cached, it's only marginally faster than
system.text() ID lookup, which is not cached and about 1/3 the speed of the
other methods of index to string lookup (in the MSVC-Windows build).