Implement double-height.
This is so gross... outputting any character depends on every
character above it, so it must be checked every time.  If the state
was ste up differently, this could be avoided, but it's not, so it
can't.

For v2, I'm going to want to store some per-line data in a separate
struct so I can track this.