Don't check recvline() timeout *before* checking if there's data to receive
This was a regression: Socket.recvline() used to not care what the timeout
duration was so long as there were bytes to receive.

Also, remove the ".0" from timeout values in documented mehtods that don't
(any longer) accept floating point timeout durations. We used to support
fractional seconds for some of these methods, and that was implied by using
the floating point default values, but that's no longer the case. poll()
still accepts a floating point timeout.