Log a build warning if building for Linux without OFD lock support
OFD locks are needed on Linux for appropriate multi-threaded shared file
access (using fcntl record locks to prevent corruption), so log a warning if
building for Linux without that support.

lock() now mimics DOS/Windows again: the result lock is an "all access" lock
regardless of what mode the file was open in. I'm not sure why this change was
made (commit 11b73134563ce26), but I don't think it was necessary or
appropriate (though I can't think of any immediate negative effects). At
minimum it makes the code a little more understandable and eliminates an
extra call to fcntl().