Support multiple filenames in subjects of file-attach netmails being packed
When writing the attached file path/names to the FLO file, parse each
filename out of the subject, accordingly. This should fix issue #824.

There's an inconsistent support in SBBSecho for FileLists in the subjects of
netmail messages. In a couple of places, single-space-separated filelists are
fine (but not commas or multiple space, even though they are allowed per
FTS-1).

This particular fix is just for the packing of netmail messages and the
appending the attached file(s) to the FLO file. Those other places (e.g.
.REQ file creation) will need their own fixes for proper filelist support.