Add a 'misc' field for System/Fixed Events allow disabling and "Native"
Configured in SCFG->External Programs->Fixed Events and the Node Daily Events
configured in SCFG->Nodes, now can be disabled (without just clearing the
command-line) and toggle-on "Native", so the program doesn't have to be added
to the dreaded "Native Program List" in order to *not* be treated as a 16-bit
DOS program (the default).

Eventually I'll allow multiple of each time of system fixed event, but this
is an improvement for now at least.