When processing AreaRequest with -R (rescan), rescan existing echoes too
If the body of the request mesages message contains a list of area-tags,
even if we already have the node linked to those echoes, do the rescan for
each listed echo. Previously, only newly-added echoes would be rescanned
via this method.

Apparently the D'bridge built-in method of generating AreaManager requests
uses this syntax (per Ward Dossche).

Also, when rescanning due to AreaManager rescan request and generating a
response netmail message, include the total number of messages exported as a
result.

Increase SBBSecho version number to v3.22 (Please upgrade, Fernando) :-)