Fix bug reported by DigitalMan
It looks like this has been a long-standing UIFC bug (pre-3.0)

If a uifc list that can scroll is showing the bottom-most line,
and an option that is not the last is deleted, and both *cur and
*bar are left untouched, *bar would be adjusted to be too high,
resulting in various highlight bar corruption issues.

This commit just rewrites that logic to be comprehensible, and
ensures that if bar is set such that the list would end before the
bottom of the window, it gets moved up.  This makes delete do the
right thing (didn't test with insert, that may do weird things).

Also, since I figured out a good definition of what bar actually
represtens, add a comment to that effect.