Overhaul the saved config.
Rather than a single file with config for all subs, use a separate
file for each sub.  This makes clients that do parallel connections
to different mailboxes work much better, and make loading and saving
the config for a mailbox much faster.

The caching stuff is removed to simplify things, and hopefully won't
need to come back... we should just not save config if it doesn't
change.

Only issue I'm still aware of is that 99% of IMAP clients just can't
deal with slashes in mailbox paths.  Like... at all.  The protocol
supports them, but every client seems to have a different way of
breaking with them.