Leave the user/*.ftp lock file open for better protection
This appears to be the only way to 100% provide mutual exclusivity between
Windows and Linux client sharing a Samba file system.

Also, make fmutex_close() take a pointer to the file descriptor so that we can
better manage the resource (i.e. don't close a descriptor that's already been
closed and possibly valid in another thread).