Fix the regex in the parsing of the "References" header field
We only want one Message-ID (we'll use the last, since that appears to be the
message being replied to based on SLRN behavior). This fixes reply linkage
when posing from newsreaders that put multiple message-IDs in the
references header field. This is required for SBBSecho to do its magic
creation of the FTN REPLY kludge during export.

This is the fix that Accession was looking for, I think.