Support AreaMgr request messages with -L (list) or -Q (query) but no body
I noticed through code review that these kinds of messages would be rejected
as "No commands to process.". We just needed to increment the 'cmd' counter
when parsing these options to be sure we don't think that no commands were
in fact request. -R (rescan) is not by itself a command since it requires
listed area-tags to have any effect.