Replace almost all references to AreaFix with AreaMgr or AreaManager
"AreaFix" is a proper product name. Though it's been defunct and unsupported
for decades, we still should not confuse sysops and historians: the general
term for what AreaFix did/does is Area Management, aka Conference Management.

The sbbsecho.ini key names are unchanged and the destination netmail address
for AreaManager requests can still be "AreaFix" (though "SBBSecho" is still
supported as well).