Create/use SysopPageNotification text.dat string
to notify sysop (user #1) that they were paged. Previously, this was
a hard-coded string. The new string in text.dat is the same, but adds
2 BELs (CTRL-G chars) and is now customizable.

Created an sbbs_t::notify() method to make this easier to use and that
method supports @-code expansion, so this text.dat string supports @-codes
(which I used).

This addresses the feature request in issue #828 (beep the sysop in the
telegram/short message, when they're paged).