Fix up the Prestel "Engineering Test Page" rendering
This was due to two main bugs... the simple one was a type where
the cutoff between separated and contiguous mosaics was incorrect,
which caused the first four separated mosaics to not be displayed.

The second was due to a number of bugs in the Hold Mosaic
implementation.

This fixes the first half of issue 172.