Just fiddling with C23...
At the same time, playing with testing in C... started at the
coverage test level and working up to see how painful it really is.

This is just a linked-list queue thing the sort that every C hacker
has implemented countless times... but if it ends up as "good", it
could be used in SyncTERM in the future.