Don't use flock() in sopen() since it ends up using non-OFD fcntl() locks
When OFD locks are available, that's what we should be using.
Otherwise, we suffer the horrible behavior of POSIX file/region locks and
a subsequent open/close of the file releases any/all locks on it.

This is currently in an !BSD block, which appears to include macOS, but
macOS *does* support OFD locks, so I'll be fixing that here shortly.