Start of DeuceSSH
The intent is that this will be a library implementation of SSH
which allows simple extension of the base protocol, and exposes all
defined features of the protocol (including ones nobody uses).

This will support both client and server, with initial focus on the
client implementation for use in SyncTERM.

The goal of the core library is to only support the most secure
standard methods.  When there are multiple methods that are not
clearly of differing levels of security, the most common two will
be implemented (ie: RSA and Ed25519).

Additionally, care will be taken during development to allow for
the OpenSSH Encrypt-then-MAC MAC modes in the expectation that
someone will standardize an ETM mode at some point, and I expect
to write an extension to support it soon after DeuceSSH is
complete.

Finally, if I cannot mitigate Terrapin without non-standard
extensions, the OpenSSH "strict KEX" extension will be supported
until there's a standard defined that solves the issue.