De-duplicate repeated errors logged via sbbs_t::errormsg() by reducing severity
... from ERROR to WARNING, so the repeats won't be logged to error.log file
and won't (normally) be sent as notifications (e.g. emails) to the sysop. The
duplicates are identified as from the same source file and line number as the
immediately previous logged error within the last 12 hours. String comparison
doesn't really work for deduplication since if you look closely, the errors
usually are *not* exact duplicates (i.e. there's a node number or a socket
descriptor or something that uniquely identifies the user/client/session).
Repeated errors don't increment the node's error counter either.

This does not de-duplicate errors logged via other means (e.g. direct calls to
log()/lputs(),lprintf() with a severity of LOG_ERR or higher), but will
solve the majority of duplicate errors that can be logged from the terminal
server.

So this at least partially addresses issue #619.