Parse console.inkey() timeout parameter as an *unsigned* integer
A negative timeout value doesn't make sense and this will extend the legal
(positive) range of this parameter value.

Hopefully this will help eliminate these errors I get often:
  dd_lightbar_menu.js line 4041: Error: can't convert inputTimeoutMS to an
  integer

But if there's some way that this parameter value is even out of the range of
an unsigned 32-bit value (4294967295) then this error could still happen.
Unfortunately, the JS exception doesn't report the value that failed to
convert.