A differnt 'fix' to link_area()
So, (sizeof *addr) now looks right got me (this should've been the same as
sizeof(fidoaddr_t)), so the change in commit e2476ebbaa doesn't actually look
like a fix, but maybee a regression.

I need to test some area manager link additions and get to the bottom of this,
but this particular change was not a fix (and caught/reported by Coverity -
CID 515673).