Fix the areas.bbs AreaMgr addition bug created in commit 813072cd450c2
The newly created/used link_aera() function was copying the address of the
passed fidoaddr_t pointer rather than the value (fidoaddr) pointed to.
This is the actual fix for the problem reported by Tom Moore (1:135/205)
- tested and validated. It only immediately impacted areas.bbs use
(not areas.ini), since we wrote a stringified address directly to the
areas.ini file "links" key. This was not an allocation issue but a memcpy()
issue.

Other unrelated changes:
- Log a debug-level message when ignoring a "NO-ECHO" message (subject "NE:*")
- Log a notice-level message when ignoring a message due to GATEWAY VIOLATION
  (gating a message between net-types when not explicitly configured to do so)