Add RELOAD sysop command to reload a (JS) command shell without logoff/on
We have to use load() (rather than js.exec) to invoke str_cmds.js so that an
exit() will actually exit. Since load() automatically does the mods vs exec
directory search-dance, that simplifies the code in default.js a little.
I'm not sure exactly why I originally chose to use js.exec() over load() for
invoking str_cmds.js, but for this feature, we need load() so let's go with
that for now.

I did encounter an issue (issue #840) while originally trying to make this
work with the original code that called js.exec(), but just punted and went
with load() instead. Perhaps if we fix issue #840, we can revert default.js
back to using js.exec() (but why we would need/want to, I'm not sure).