Fix bug in recursive use of delfiles(), not deleting all subdirs/files
Looks like this feature (commit d661427e13) never really worked correctly
since it counted the files removed from each sub-dir and then stopped deleting
when the count reached the number of files in the base directory. This was
done to accommodate the 'keep' feature (part of previous commits).
So make 'keep' check conditional on it being non-zero and just don't ever use
a non-zero keep value with a recursive delete and we should be good! :-)

This fixes issue #841