Fix issue when receiving node messages while using down-arrow at pause prompt
Hitting down-arrow key at a pause prompt normally displaye just one more line
of the display text/file, but if you received a node message/telegram/notice
after hitting down arrow, you'd get a screen full of text instead of just a
single (one more) line, as you wanted.

This looks to be because of the anti-recursive protection implemented in
pause() - when it calls nodesync() after the key press, that displays
node/user messages (if there are any) and if pause is called as a result
(e.g. because the line counter was already set to cause a pause after the
next line of output), it'd do nothing since that would be recursive. The fix
is to simply set the line counter as a result of the down-arrow key press
*after* the call to nodesync(), which might display multiple lines, but I
think that's fine.