Document new support for JPEG XL
An alternative to using PPM, relies on libjxl from the JPEG.
JPEG XL is a superior modern format that supports both lossy and
lossless compression and does an excellent job at both.

The format also supports multiple frames (ie: animation), so can
be used for Cool Stuff in the future.

As an example, a large image that is 130MB in PPM format is only
6.8MB in JPEG XL format using the default quality settings of
ImageMagick (whatever they are).