Fix GTK stuff for macOS
Remove the -E linker flag.
This should not be needed anymore, and isn't supported on macOS.

Have shared libraries include their full path.
This allows linked dylibs to work from where they were built, so
as long as you don't build the binaries on a CI machine, then try
to run them on a users machine (*cough*), it'll work out.

Use the correct rpath argument format on macOS.
It uses -rpath, not --rpath.