In X11 mode, support arbitrary character entry.
You still cannot enter a NUL.

Holding down ALT and entering an 8-bit decimal number on the keypad
starting with a digit from 1-9 will pass that byte to the input
buffer as entered.

If you prefix the decimal number with the keypad zero, you can enter
a Unicode codepoint to be translated and sent (if it translates).