Give SSH the same treatment TelnetS just got.
Use an atomic that's checked in the while loop in the io threads,
for SSH, we needed to add a new one since ssh_active indicates the
cryptlib session is active, so use ssh_complete to indicate that
a CRYPT_ERROR_COMPLETE (or other final message) occured.

This should fix issue 174.